@@ -2,8 +2,9 @@ package main
 
 import (
 	"context"
+	"errors"
 	"fmt"
-	"log"
+	"log/slog"
 	"net/http"
 	"os"
 
@@ -20,14 +21,123 @@ import (
 )
 
 var (
-	serviceName      = getEnvOrDefault("SERVICE_NAME", "product-service")
-	collectorURL     = getEnvOrDefault("OPENOBSERVE_URL", "http://36.64.55.158:5080")
-	EnvOtlpAuthToken = "OPENOBSERVE_AUTH_TOKEN"
-	DefaultAuthToken = ""
-	EnvPort          = "PORT"
-	DefaultPort      = "8086"
-	tracer           = otel.Tracer(serviceName)
+	serviceApp  = getEnvOrDefault("APPLICATION", "ecommerce")
+	serviceEnv  = getEnvOrDefault("ENVIRONMENT", "development")
+	collectorURL = getEnvOrDefault("OTLP_URL", "http://tempo:4318/v1/traces")
+
+	serviceName  = getEnvOrDefault("SERVICE_NAME", "product-service")
+	EnvPort      = "PORT"
+	DefaultPort  = "8086"
+	tracer       = otel.Tracer(serviceName)
 )
+
+func convertSlogAttrsToAPMAttrsFromSlice(slogAttrs []slog.Attr) []attribute.KeyValue {
+    spanAttributes := make([]attribute.KeyValue, 0, len(slogAttrs))
+    for _, attr := range slogAttrs {
+        spanAttributes = append(spanAttributes, attribute.String(attr.Key, attr.Value.String()))
+    }
+    return spanAttributes
+}
+
+// APMHandler is a custom slog.Handler that adds OpenTelemetry trace/span IDs
+// and records errors on the active span.
+type APMHandler struct {
+	slog.Handler
+	attrs []slog.Attr
+}
+
+// NewAPMHandler creates a new APMHandler that wraps a base slog.Handler.
+func NewAPMHandler(baseHandler slog.Handler) *APMHandler {
+	return &APMHandler{
+		Handler: baseHandler,
+	}
+}
+
+// Handle implements slog.Handler.Handle. It adds trace/span IDs and records errors.
+func (h *APMHandler) Handle(ctx context.Context, r slog.Record) error {
+	span := trace.SpanFromContext(ctx)
+	if span.IsRecording() {
+		spanCtx := span.SpanContext()
+		if spanCtx.HasTraceID() {
+			r.AddAttrs(slog.String("trace.id", spanCtx.TraceID().String()))
+		}
+		if spanCtx.HasSpanID() {
+			r.AddAttrs(slog.String("span.id", spanCtx.SpanID().String()))
+		}
+
+		// Collect ALL attributes: those from logger.With AND those from the current log call
+		allAttrs := make([]slog.Attr, 0, len(h.attrs)+r.NumAttrs())
+		allAttrs = append(allAttrs, h.attrs...) // 1. Start with handler's accumulated WithAttrs
+
+		// 2. Add attributes from the current slog.Record itself
+		r.Attrs(func(attr slog.Attr) bool {
+			allAttrs = append(allAttrs, attr)
+			return true
+		})		
+
+		// If it's an error level, record the error on the span
+		if r.Level == slog.LevelError {
+			// Check if an error was explicitly passed as an attribute using slog.Any("error", err)
+			var loggedErr error
+			r.Attrs(func(attr slog.Attr) bool {
+				if attr.Key == "error" { // We'll use "error" as the key for explicit errors
+					if errVal, ok := attr.Value.Any().(error); ok {
+						loggedErr = errVal
+						return false // Found it, stop iterating
+					}
+				}
+				return true // Continue iterating
+			})
+
+			if loggedErr != nil {
+				span.RecordError(loggedErr, trace.WithAttributes(
+					attribute.String("event", "log_error"),
+					attribute.String("message", r.Message),
+				))
+				span.SetStatus(codes.Error, loggedErr.Error())
+			} else {
+				// Record the log message itself as an error if no specific error object
+				span.RecordError(errors.New(r.Message), trace.WithAttributes(
+					attribute.String("event", "log_error"),
+				))
+				span.SetStatus(codes.Error, r.Message)
+			}
+		} else if r.Level == slog.LevelInfo || r.Level == slog.LevelWarn {
+			// Convert slog attributes to OpenTelemetry attributes using the helper function
+			APMAttrs := convertSlogAttrsToAPMAttrsFromSlice(allAttrs)
+			// Add an event to the OpenTelemetry span with the record's message and converted attributes
+			span.AddEvent(r.Message, trace.WithAttributes(APMAttrs...))
+		}
+	}
+
+	// Always call the base handler to actually output the log record
+	return h.Handler.Handle(ctx, r)
+}
+
+
+// WithAttrs implements slog.Handler.WithAttrs.
+func (h *APMHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
+    newAttrs := make([]slog.Attr, 0, len(h.attrs)+len(attrs))
+    newAttrs = append(newAttrs, h.attrs...) // Include existing attrs
+    newAttrs = append(newAttrs, attrs...)    // Add new attrs
+
+    return &APMHandler{
+        Handler: h.Handler.WithAttrs(attrs), // Still pass new attrs to the wrapped handler
+        attrs:   newAttrs,                   // Store the combined attrs in the new APMHandler
+    }
+}
+
+// WithGroup implements slog.Handler.WithGroup.
+func (h *APMHandler) WithGroup(name string) slog.Handler {
+	return NewAPMHandler(h.Handler.WithGroup(name))
+}
+
+// Enabled implements slog.Handler.Enabled.
+func (h *APMHandler) Enabled(ctx context.Context, level slog.Level) bool {
+	return h.Handler.Enabled(ctx, level)
+}
+
+
 
 // getTraceID extracts the trace ID from the current span context
 func getTraceID(ctx context.Context) string {
@@ -65,12 +175,7 @@ func initTracerProvider(serviceName string, collectorURL string) (*sdktrace.Trac
 
 	// Buat OTLP HTTP Exporter
 	client := otlptracehttp.NewClient(
-		otlptracehttp.WithEndpoint(collectorURL),
-		otlptracehttp.WithInsecure(),                        // Gunakan ini jika OpenObserve tidak menggunakan HTTPS
-		otlptracehttp.WithURLPath("/api/default/v1/traces"), // Path untuk OpenObserve
-		otlptracehttp.WithHeaders(map[string]string{
-			"Authorization": "Basic " + getEnvOrDefault(EnvOtlpAuthToken, DefaultAuthToken),
-		}),
+		otlptracehttp.WithEndpointURL(collectorURL),
 	)
 	exporter, err := otlptrace.New(ctx, client)
 	if err != nil {
@@ -83,8 +188,8 @@ func initTracerProvider(serviceName string, collectorURL string) (*sdktrace.Trac
 		sdktrace.WithResource(resource.NewWithAttributes(
 			semconv.SchemaURL,
 			semconv.ServiceNameKey.String(serviceName),
-			attribute.String("application", "ecommerce"),
-			attribute.String("environment", "development"),
+			attribute.String("application", serviceApp),
+			attribute.String("environment", serviceEnv),
 		)),
 		sdktrace.WithSampler(sdktrace.AlwaysSample()), // Selalu sample semua trace
 	)
@@ -102,60 +207,108 @@ func initTracerProvider(serviceName string, collectorURL string) (*sdktrace.Trac
 }
 
 func main() {
+	// 1. Initialize Tracer Provider
 	tp, err := initTracerProvider(serviceName, collectorURL)
 	if err != nil {
-		log.Fatalf("[system] Gagal menginisialisasi TracerProvider: %v", err)
+		// Use slog.Error and then os.Exit for fatal errors, as slog doesn't have a direct Fatal
+		slog.Default().Error("Gagal menginisialisasi TracerProvider",
+			"context", "main",
+			slog.Any("error", err), // Use slog.Any for errors
+		)
+		os.Exit(1)
 	}
 	defer func() {
 		if err := tp.Shutdown(context.Background()); err != nil {
-			log.Printf("[system] Error saat shutdown TracerProvider: %v", err)
+			slog.Default().Error("Error saat shutdown TracerProvider",
+				"context", "main",
+				slog.Any("error", err),
+			)
 		}
 	}()
+
+	// 2. Configure slog logger
+	// Set up a base JSON handler for structured output to stdout
+	jsonHandler := slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{
+		AddSource: true, // Add file and line number to logs
+		Level:     slog.LevelDebug,
+	})
+
+	// Wrap the base handler with our APMHandler to inject trace/span IDs and handle errors
+	logger := slog.New(NewAPMHandler(jsonHandler))
+
+	// Set the default logger for convenience if you want to use slog.Info etc. directly
+	slog.SetDefault(logger)
 
 	repo := NewProductRepository()
 	service := NewProductService(repo)
 
+	ctx := context.Background()
 	http.HandleFunc("/product", func(w http.ResponseWriter, r *http.Request) {
-		handleProduct(w, r, service)
+		ctx = otel.GetTextMapPropagator().Extract(r.Context(), 
+			propagation.HeaderCarrier(r.Header))
+		ctx, span := tracer.Start(ctx, "/product",
+			trace.WithAttributes(
+				attribute.String("http.method", r.Method),
+				attribute.String("http.url", r.URL.String()),
+			))
+		defer span.End()
+		// Pass the logger with context. Context not needed if you use AddSource: true
+		// handleProduct(ctx, logger.With("context", "handleProduct"), w, r, service)
+		handleProduct(ctx, logger, w, r, service)
 	})
 
 	port := getEnvOrDefault(EnvPort, DefaultPort)
 	addr := ":" + port
-	log.Printf("[%s] %s berjalan di %s", "system", serviceName, addr)
-	log.Fatal(http.ListenAndServe(addr, nil))
+	logger.With(
+		"context", "main",
+		"serviceName", serviceName,
+		"address", addr,
+	).Info("berjalan")
+
+	if listenErr := http.ListenAndServe(addr, nil); listenErr != nil {
+		logger.With(
+			"context", "main",
+			slog.Any("error", listenErr),
+		).Error("Server stopped with an error")
+		os.Exit(1)
+	}
 }
 
-func handleProduct(w http.ResponseWriter, r *http.Request, service ProductService) {
-	ctx := otel.GetTextMapPropagator().Extract(r.Context(), propagation.HeaderCarrier(r.Header))
-	ctx, span := tracer.Start(ctx, "handleProduct",
-		trace.WithAttributes(
-			attribute.String("http.method", r.Method),
-			attribute.String("http.url", r.URL.String()),
-		))
+func handleProduct(ctx context.Context, logger *slog.Logger, 
+	w http.ResponseWriter, r *http.Request, service ProductService) {
+	productID := r.URL.Query().Get("id")
+	ctx, span := tracer.Start(ctx, "handleProduct", trace.WithAttributes(attribute.String("product.id", productID)))
 	defer span.End()
 
-	productID := r.URL.Query().Get("id")
 	if productID == "" {
+		logger.With(
+			slog.Any("error", "Missing product ID"),
+		).ErrorContext(ctx, "Missing product ID")
 		http.Error(w, "Parameter 'id' produk diperlukan", http.StatusBadRequest)
-		span.SetStatus(codes.Error, "Missing product ID")
 		return
 	}
-	span.SetAttributes(attribute.String("product.id", productID))
 
-	traceID, spanID := getTraceSpanInfo(ctx)
-	log.Printf("[%s|%s] Product Service: Mencari info produk ID %s", traceID, spanID, productID)
+	logger.With(
+		"productID", productID,
+		).DebugContext(ctx, "Mencari info produk")
 
 	// Service Layer: Get Product Info (with trace)
-	productInfo, err := service.GetProductInfo(ctx, productID)
+	productInfo, err := service.GetProductInfo(ctx, logger, productID)
 	if err != nil {
-		log.Printf("[%s|%s] Product Service: Error processing product ID %s: %v", traceID, spanID, productID, err)
-		span.RecordError(err)
-		span.SetStatus(codes.Error, "Failed to fetch product info")
+		logger.With(
+			"productID", productID,
+			slog.Any("error", err), // Use slog.Any to pass the error object
+		).ErrorContext(ctx, "Failed to fetch product info")
 		http.Error(w, "Gagal mendapatkan info produk", http.StatusInternalServerError)
 		return
 	}
-	span.AddEvent("Product data fetched", trace.WithAttributes(attribute.String("product.info", productInfo)))
+	logger.With(
+		"productID", productID,
+		"productInfo", productInfo,
+		).InfoContext(ctx, "Product info fetched")
 
-	log.Printf("[%s|%s] Product Service: Successfully processed request for product ID %s", traceID, spanID, productID)
+	logger.With(
+		"productID", productID,
+		).DebugContext(ctx, "Successfully processed request")
 	fmt.Fprint(w, productInfo)
 }
