@@ -2,34 +2,42 @@ package main
 
 import (
 	"context"
-	"log"
+	"log/slog"
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
 )
 
 type ProductService interface {
-	GetProductInfo(ctx context.Context, productID string) (string, error)
+	GetProductInfo(ctx context.Context, logger *slog.Logger, productID string) (string, error)
 }
 
 type productServiceImpl struct {
 	repo ProductRepository
 }
 
-func (s *productServiceImpl) GetProductInfo(ctx context.Context, productID string) (string, error) {
+func (s *productServiceImpl) GetProductInfo(ctx context.Context, logger *slog.Logger, 
+	productID string) (string, error) {
+
 	ctx, span := tracer.Start(ctx, "ProductService.GetProductInfo", trace.WithAttributes(attribute.String("product.id", productID)))
 	defer span.End()
 
-	traceID, spanID := getTraceSpanInfo(ctx)
-	log.Printf("[%s|%s] ProductService: Processing request for product ID %s", traceID, spanID, productID)
+	logger.With(
+		"productID", productID,
+	).DebugContext(ctx, "Processing request")
 
-	productInfo, err := s.repo.FetchProduct(ctx, productID)
+	productInfo, err := s.repo.FetchProduct(ctx, logger, productID)
 	if err != nil {
-		log.Printf("[%s|%s] ProductService: Error fetching product %s: %v", traceID, spanID, productID, err)
+		logger.With(
+			"productID", productID,
+		).ErrorContext(ctx, "Error fetching product")
 		return "", err
 	}
 
-	log.Printf("[%s|%s] ProductService: Successfully retrieved product info for %s", traceID, spanID, productID)
+	logger.With(
+		"productID", productID,
+		"productInfo", productInfo,
+	).InfoContext(ctx, "Successfully retrieved product info")
 	return productInfo, nil
 }
 
